Time: 4 ms (95.43%) | Memory: 19.3 MB (71.53%) - LeetSync

+++ 45-jump-game-ii/jump-game-ii.cpp
@@ -3,7 +3,7 @@ public:
     int jump(vector<int>& nums) {
         int n=nums.size();
         int l=0,r=0;
-        long long jumps=0;
+        int jumps=0;
         while(r<n-1){
             int farthest=0;
             for(int i=l;i<=r;i++){
